open file attachment support
