+++ app/pubspec.yaml
@@ -90,6 +90,7 @@ dependencies:
   dart_date: ^1.3.2
   path: ^1.8.3
   url_launcher: ^6.1.10
+  open_filex: ^4.4.0
   file_picker: ^6.1.1
   emoji_picker_flutter: ^2.1.1
   connectivity_plus: ^5.0.2
